requirements.txt
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0
torchdiffeq>=0.2.3
pyyaml>=6.0
tqdm>=4.65.0
